@@ -11,19 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/architect": "^0.1802.2",
-    "@angular/animations": "^18.0.5",
-    "@angular/cdk": "^18",
-    "@angular/common": "^18.0.5",
-    "@angular/compiler": "^18.0.5",
-    "@angular/core": "^18.0.5",
-    "@angular/forms": "^18.0.5",
-    "@angular/localize": "^18.0.5",
-    "@angular/platform-browser": "^18.0.5",
-    "@angular/platform-browser-dynamic": "^18.0.5",
-    "@angular/router": "^18.0.5",
-    "@asymmetrik/ngx-leaflet": "^18",
-    "@ng-bootstrap/ng-bootstrap": "^17",
+    "@angular-devkit/architect": "^0.1900.6",
+    "@angular/animations": "^19.0.5",
+    "@angular/cdk": "^19",
+    "@angular/common": "^19.0.5",
+    "@angular/compiler": "^19.0.5",
+    "@angular/core": "^19.0.5",
+    "@angular/forms": "^19.0.5",
+    "@angular/localize": "^19.0.5",
+    "@angular/platform-browser": "^19.0.5",
+    "@angular/platform-browser-dynamic": "^19.0.5",
+    "@angular/router": "^19.0.5",
+    "@bluehalo/ngx-leaflet": "^19",
+    "@ng-bootstrap/ng-bootstrap": "^18",
     "@ng-bootstrap/schematics": "^2.0.0-alpha.1",
     "@ngx-grpc/common": "^3.1.1",
     "@ngx-grpc/core": "^3.1.1",
@@ -39,49 +39,49 @@
     "core-js": "^3.28.0",
     "escape-string-regexp": "^5.0.0",
     "google-proto-files": "^4.2.0",
-    "keycloak-angular": "^16",
+    "keycloak-angular": "^19",
     "keycloak-js": "^25 || ^26",
     "leaflet": "^1.9.3",
-    "ng-recaptcha-2": "^14",
+    "ng-recaptcha-2": "^15",
     "ng2-charts": "^6 || ^7",
     "ngx-autosize": "^2.0.4",
     "ngx-pipes": "^3.2.0",
-    "npm": "^10.2.1",
+    "npm": "^10 || ^11",
     "sprintf-js": "^1.1.2",
     "tslib": "^2.5.0",
-    "typescript": ">=5.4.0 && <5.6.0",
+    "typescript": ">=5.4.0 && <5.8.0",
     "url-parse": "^1.5.10",
-    "zone.js": "^0.14.4 || ^0.15"
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "^18",
-    "@angular-eslint/eslint-plugin": "^18",
-    "@angular-eslint/eslint-plugin-template": "^18",
-    "@angular-eslint/schematics": "^18",
-    "@angular-eslint/template-parser": "^18",
-    "@angular/build": "^18.2.9",
-    "@angular/cli": "^18.0.6",
-    "@angular/compiler-cli": "^18.0.5",
-    "@angular/language-service": "^18.0.5",
+    "@angular-eslint/builder": "^19.0.2",
+    "@angular-eslint/eslint-plugin": "^19",
+    "@angular-eslint/eslint-plugin-template": "^19",
+    "@angular-eslint/schematics": "^19",
+    "@angular-eslint/template-parser": "^19",
+    "@angular/build": "^19.0.6",
+    "@angular/cli": "^19.0.6",
+    "@angular/compiler-cli": "^19.0.5",
+    "@angular/language-service": "^19.0.5",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.17.0",
     "@ngx-grpc/protoc-gen-ng": "^3.1.1",
-    "@types/diff": "^5.2 || ^6",
+    "@types/diff": "^6",
     "@types/google-protobuf": "^3.15.6",
     "@types/jquery": "^3.5.16",
     "@types/leaflet": "^1.9.1",
     "@types/node": "^22",
     "@types/showdown": "^2.0.6",
     "@types/sprintf-js": "^1.1.4",
-    "@typescript-eslint/eslint-plugin": "^7 || ^8",
-    "@typescript-eslint/parser": "^7 || ^8",
-    "eslint": "^8.45.0 || ^9",
+    "@typescript-eslint/eslint-plugin": "^8",
+    "@typescript-eslint/parser": "^8",
+    "eslint": "^9",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-depend": "^0.11 || ^0.12",
+    "eslint-plugin-depend": "^0.12",
     "eslint-plugin-json": "^4",
-    "eslint-plugin-perfectionist": "^2.2.0 || ^3 || ^4",
+    "eslint-plugin-perfectionist": "^4",
     "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-rxjs": "^5.0.2",
-    "eslint-plugin-rxjs-angular": "^2.0.0",
-    "eslint-plugin-sonarjs": "^2 || ^3",
+    "eslint-plugin-sonarjs": "^3",
     "google-protobuf": "^3.21.2",
     "grpc-web": "^1.4.2",
     "html-comment-regex": "^1.1.2",
