@@ -9,17 +9,16 @@ import {ContactsService} from '@services/contacts';
 import {LanguageService} from '@services/language';
 import {PageEnvService} from '@services/page-env.service';
 import {TimeAgoPipe} from '@utils/time-ago.pipe';
-import {KeycloakService} from 'keycloak-angular';
 import {BehaviorSubject, EMPTY, Observable} from 'rxjs';
 import {catchError, map, switchMap} from 'rxjs/operators';
 
 import {ToastsService} from '../../toasts/toasts.service';
 import {UserComponent} from '../../user/user/user.component';
+import Keycloak from 'keycloak-js';
 
 @Component({
   imports: [RouterLink, UserComponent, NgbTooltip, AsyncPipe, DatePipe, TimeAgoPipe],
   selector: 'app-account-contacts',
-  standalone: true,
   templateUrl: './contacts.component.html',
 })
 export class AccountContactsComponent implements OnInit {
@@ -29,7 +28,7 @@ export class AccountContactsComponent implements OnInit {
   private readonly auth = inject(AuthService);
   private readonly contacts = inject(ContactsClient);
   private readonly languageService = inject(LanguageService);
-  private readonly keycloak = inject(KeycloakService);
+  private readonly keycloak = inject(Keycloak);
 
   private readonly reload$ = new BehaviorSubject<void>(void 0);
 
