@@ -1,5 +1,5 @@
 import {AsyncPipe, NgClass} from '@angular/common';
-import {Component, inject, Renderer2} from '@angular/core';
+import {Component, effect, inject, Renderer2} from '@angular/core';
 import {NavigationStart, Router, RouterLink, RouterLinkActive, RouterOutlet} from '@angular/router';
 import {environment} from '@environment/environment';
 import {APIUser, ItemFields, ItemListOptions, ItemType, ListItemsRequest} from '@grpc/spec.pb';
@@ -19,13 +19,14 @@ import {MessageService} from '@services/message';
 import {LayoutParams, PageEnvService} from '@services/page-env.service';
 import {MarkdownComponent} from '@utils/markdown/markdown.component';
 import {Angulartics2GoogleAnalytics} from 'angulartics2';
-import {KeycloakService} from 'keycloak-angular';
 import {Observable} from 'rxjs';
 import {shareReplay} from 'rxjs/operators';
 
 import {MenuComponent} from './moder/menu/menu/menu.component';
 import {ContainerComponent} from './toasts/container/container.component';
 import {UsersOnlineComponent} from './users/online/online.component';
+import Keycloak from 'keycloak-js';
+import {KEYCLOAK_EVENT_SIGNAL, KeycloakEventType} from 'keycloak-angular';
 
 @Component({
   imports: [
@@ -44,7 +45,6 @@ import {UsersOnlineComponent} from './users/online/online.component';
     AsyncPipe,
   ],
   selector: 'app-root',
-  standalone: true,
   styleUrls: ['./app.component.scss'],
   templateUrl: './app.component.html',
 })
@@ -57,7 +57,7 @@ export class AppComponent {
   private readonly languageService = inject(LanguageService);
   private readonly modalService = inject(NgbModal);
   private readonly renderer = inject(Renderer2);
-  private readonly keycloak = inject(KeycloakService);
+  private readonly keycloak = inject(Keycloak);
   private readonly itemsClient = inject(ItemsClient);
 
   protected languages: Language[] = environment.languages;
@@ -84,6 +84,8 @@ export class AppComponent {
   protected language: string = this.languageService.language;
   protected urlPath = '/';
   protected isNavbarCollapsed = true;
+
+  private readonly keycloakSignal = inject(KEYCLOAK_EVENT_SIGNAL);
 
   constructor() {
     const angulartics2GoogleAnalytics = inject(Angulartics2GoogleAnalytics);
@@ -114,6 +116,16 @@ export class AppComponent {
     if (environment.production) {
       angulartics2GoogleAnalytics.startTracking();
     }
+
+    effect(() => {
+      const keycloakEvent = this.keycloakSignal();
+
+      if (keycloakEvent.type === KeycloakEventType.TokenExpired) {
+        this.keycloak.updateToken().catch((error) => {
+          console.error('Failed to refresh token', error);
+        });
+      }
+    });
   }
 
   protected doLogin() {
