@@ -6,7 +6,6 @@ import {AuthService} from '@services/auth.service';
 import {LanguageService} from '@services/language';
 import {PageEnvService} from '@services/page-env.service';
 import {APIPictureGetResponse, PictureService} from '@services/picture';
-import {KeycloakService} from 'keycloak-angular';
 import {combineLatest, EMPTY, Observable, of} from 'rxjs';
 import {catchError, debounceTime, distinctUntilChanged, map, switchMap} from 'rxjs/operators';
 
@@ -14,6 +13,7 @@ import {PaginatorComponent} from '../paginator/paginator/paginator.component';
 import {ThumbnailComponent} from '../thumbnail/thumbnail/thumbnail.component';
 import {ToastsService} from '../toasts/toasts.service';
 import {APIInbox, InboxService} from './inbox.service';
+import Keycloak from 'keycloak-js';
 
 const ALL_BRANDS = 'all';
 
@@ -26,7 +26,6 @@ interface Inbox {
 @Component({
   imports: [RouterLink, FormsModule, ThumbnailComponent, PaginatorComponent, AsyncPipe, DatePipe],
   selector: 'app-inbox',
-  standalone: true,
   templateUrl: './inbox.component.html',
 })
 export class InboxComponent implements OnInit {
@@ -34,7 +33,7 @@ export class InboxComponent implements OnInit {
   private readonly auth = inject(AuthService);
   private readonly route = inject(ActivatedRoute);
   private readonly languageService = inject(LanguageService);
-  private readonly keycloak = inject(KeycloakService);
+  private readonly keycloak = inject(Keycloak);
   private readonly pictureService = inject(PictureService);
   private readonly inboxService = inject(InboxService);
   private readonly pageEnv = inject(PageEnvService);
