@@ -1,6 +1,6 @@
 import {Component, ComponentRef, inject, NgZone, OnInit, ViewContainerRef} from '@angular/core';
 import {RouterLink} from '@angular/router';
-import {LeafletModule} from '@asymmetrik/ngx-leaflet';
+import {LeafletModule} from '@bluehalo/ngx-leaflet';
 import {MapGetPointsRequest, MapPoint} from '@grpc/spec.pb';
 import {MapClient} from '@grpc/spec.pbsc';
 import {PageEnvService} from '@services/page-env.service';
@@ -25,7 +25,6 @@ function createMarker(lat: number, lng: number): Marker {
 @Component({
   imports: [RouterLink, LeafletModule],
   selector: 'app-map',
-  standalone: true,
   styleUrls: ['./styles.scss'],
   templateUrl: './map.component.html',
 })
